\name{bugsex}
\docType{data}

\alias{bugsex}
\alias{Air}
\alias{Alligators}
\alias{Asia}
\alias{Beetles}
\alias{Biopsies}
\alias{BiRats}
\alias{Blockers}
\alias{Bones}
\alias{Camel}
\alias{Cervix}
\alias{Dogs}
\alias{Dugongs}
\alias{Dyes}
\alias{Elevation}
\alias{Endo}
\alias{Epil}
\alias{Equiv}
\alias{Eyes}
\alias{Eyetracking}
\alias{Forest}
\alias{Hearts}
\alias{HepatitisME}
\alias{Hepatitis}
\alias{Hips2}
\alias{Hips3}
\alias{Hips4}
\alias{Huddersfield}
\alias{Ice}
\alias{Inhalers}
\alias{Jama}
\alias{Jaws}
\alias{Kidney}
\alias{Leukfr}
\alias{Leuk}
\alias{LHA}
\alias{Line}
\alias{Lsat}
\alias{Magnesium}
\alias{Mice}
\alias{MVCAR}
\alias{Otrees}
\alias{OtreesMVN}
\alias{Oxford}
\alias{Pigs}
\alias{Pigweights}
\alias{Pines}
\alias{Pollution}
\alias{Pumps}
\alias{Rats}
\alias{Salm}
\alias{Schools}
\alias{Scotland1}
\alias{Scotland}
\alias{Seeds}
\alias{Shared}
\alias{Sixcomp}
\alias{Stacks}
\alias{Stagnantc}
\alias{StVeit}
\alias{Surgical}
\alias{Surgicalrand}
\alias{t-df}

\title{Standard BUGS examples}
\description{
This data set gives the names of 62 standard BUGS examples and
indicates the BUGS engines under which the examples can be run.
}
\format{
  A data frame with 62 rows and 4 columns:
  \describe{
    \item{name}{A character vector containing the names of the
      examples.
    }
    \item{WinBUGS}{Logical vector indicating whether the example can be
      run under WinBUGS.
    }
    \item{OpenBUGS}{Logical vector indicating whether the example can be
      run under OpenBUGS.
    }
    \item{JAGS}{Logical vector indicating whether the example can be
      run under JAGS.
    }
  }
}
\keyword{datasets}
\seealso{\code{\link{runExample}}}
